bin: add mac-compat lock script

+++ bin/lock.sh
@@ -1,13 +1,17 @@
-#!/bin/sh
+#!/usr/bin/env bash
 
-# take screenshot
-import -window root /tmp/screenshot.png
+if [[ `uname` == 'Darwin' ]]; then
+  /System/Library/CoreServices/"Menu Extras"/User.menu/Contents/Resources/CGSession -suspend
+else
+  # take screenshot
+  import -window root /tmp/screenshot.png
 
-# blur it
-convert /tmp/screenshot.png -blur 0x9 /tmp/screenshotblur.png
-rm /tmp/screenshot.png
+  # blur it
+  convert /tmp/screenshot.png -blur 0x9 /tmp/screenshotblur.png
+  rm /tmp/screenshot.png
 
-# lock the screen
-i3lock -i /tmp/screenshotblur.png
+  # lock the screen
+  i3lock -i /tmp/screenshotblur.png
 
-exit 0
+  exit 0
+fi
